+++ test/faker/crypto_coin_test.exs
@@ -1,5 +1,0 @@
-defmodule Faker.CryptoCoinTest do
-  use ExUnit.Case, async: true
-
-  doctest Faker.CryptoCoin
-end