+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 * Added: Category `Food` for `Armenian Language` [[@hovikman][]]
 * Added: Support for `Armenian Language` [[@hovikman][]]
 * Added: `Faker.Vehicle` [[@daytonn][]]
-* Added: `Faker.CryptoCoin` [[@thadeubrito][]]
+* Added: `Faker.BlockChain.CryptoCoin` [[@thadeubrito][]]
 
 ### Changed